requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
boto3==1.35.16
botocore==1.35.16
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
fastavro==1.9.7
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.16
langchain-cohere==0.2.4
langchain-community==0.2.16
langchain-core==0.2.39
langchain-experimental==0.0.65
langchain-text-splitters==0.2.4
langsmith==0.1.117
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-requests==2.32.0.20240907
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.11.1
